fix compilation error

+++ ForkEat/ForkEat.Core/Contracts/GetIngredientResponseWithImage.cs
@@ -21,7 +21,7 @@ namespace ForkEat.Core.Contracts
         public Guid ImageId { get; set; }
         public string Name { get; set; }
         public Guid ProductId { get; set; }
-        public uint Quantity { get; set; }
+        public double Quantity { get; set; }
         public UnitResponse Unit { get; set; }
     }
 }
